[T] created class interactor
class interactor, which interacts with board by putting or moving stones
